Reorganise into src folder
