+++ src/generateReactionToNews.js
