+++ test/app.test.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-import app from '../app.js';
+import app from '../src/app.js';
 import supertest from 'supertest';
 
 let request;
